Initial commit. Basic functionality is working. It is possible to select two time zones and see the time in both zones over a 24 hour period using a range slider.
There's a lot of unused code in this commit. But I'm holding onto it in case I need to roll back to using the Google timezone API to get timezone names, instead of using Spacetime.

Google and GitGuardian will take issue with the API key being visible. The key is restricted to accept requests only from one IP address. So although it is public, it won't be useful to anyone.