Added a new web page to try out a proper design. Looks OK so far, many changes needed still.
Made some changes to index.html to try out having the autocomplete input box appear as plain text until it is clicked on. Might use this as the primary input method in the final design, but it might not appear obvious to users, so will have to try this out.